add onHover to show local time and when the market is closing
