+++ app/src/components/common/form/date_field/index.tsx
@@ -1,9 +1,9 @@
-import moment from 'moment-timezone'
 import React from 'react'
 import DatePicker from 'react-datepicker'
 import ReactDOM from 'react-dom'
 import styled from 'styled-components'
 
+import { convertLocalToUTC, convertUTCToLocal } from '../../../../util/tools'
 import { CommonDisabledCSS } from '../common_styled'
 
 interface Props {
@@ -103,26 +103,6 @@ const CalendarPortal = (props: CalendarPortalProps) => {
   )
 }
 
-// we need to do this because the value selected by react-datepicker
-// uses the local timezone, but we want to interpret it in UTC
-const convertLocalToUTC = (date: Date): Date => {
-  const offsetMinutes = moment().utcOffset()
-  return moment(date)
-    .add(offsetMinutes, 'minutes')
-    .toDate()
-}
-
-const convertUTCToLocal = (date: Maybe<Date>): Maybe<Date> => {
-  if (!date) {
-    return date
-  }
-  const offsetMinutes = moment().utcOffset()
-
-  return moment(date)
-    .subtract(offsetMinutes, 'minutes')
-    .toDate()
-}
-
 export const DateField = (props: Props) => {
   const { disabled, minDate, name, onChange, selected, ...restProps } = props
 
@@ -137,7 +117,7 @@ export const DateField = (props: Props) => {
         calendarClassName="customCalendar"
         dateFormat="MMMM d, yyyy h:mm aa"
         disabled={disabled}
-        minDate={new Date()}
+        minDate={minDate}
         name={name}
         onChange={handleChange}
         placeholderText="Select Date"
