+++ app/src/components/market/sections/market_create/steps/items/ask_question_step.tsx
@@ -188,6 +188,7 @@ const AskQuestionStep = (props: Props) => {
             formField={
               <DateField
                 disabled={!!loadedQuestionId}
+                minDate={new Date()}
                 name="resolution"
                 onChange={handleDateChange}
                 selected={resolution}
